# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/seed/rotate:
    post:
      operationId: rotateGameSeed
      tags: [games]
      summary: Reveal the current seed and start a new one
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotateSeedRequest"
      responses:
        "200":
          description: Previous seed revealed; new seed committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RotateSeedResult"
  /games/dice/roll:
    post:
      operationId: rollDice
      tags: [games]
      summary: Place a dice bet, settled server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiceRollRequest"
      responses:
        "200":
          description: Bet settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiceRollResponse"
  /games/roulette/spin:
    post:
      operationId: spinRoulette
      tags: [games]
      summary: Place a roulette bet, settled server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouletteSpinRequest"
      responses:
        "200":
          description: Bet settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouletteSpinResponse"
  /games/slots/spin:
    post:
      operationId: spinSlots
      tags: [games]
      summary: Spin the slots, settled server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotsSpinRequest"
      responses:
        "200":
          description: Bet settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsSpinResponse"
  /games/lucky-pick/draw:
    post:
      operationId: drawLuckyPick
      tags: [games]
      summary: Place a lucky-pick bet, settled server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LuckyPickDrawRequest"
      responses:
        "200":
          description: Bet settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LuckyPickDrawResponse"
...
components:
  schemas:
...
    RotateSeedRequest:
      type: object
      properties:
        clientSeed:
          type: string
          description: Optional player-supplied seed for the new commitment. A random one is generated if omitted.
    RotateSeedResult:
      type: object
      properties:
        revealedServerSeed:
          type: string
          nullable: true
          description: The raw seed that was active until now (null if there was no prior seed to reveal).
        revealedServerSeedHash:
          type: string
          nullable: true
        newServerSeedHash:
          type: string
        clientSeed:
          type: string
      required:
        - newServerSeedHash
        - clientSeed
    DiceRollRequest:
      type: object
      properties:
        betAmount:
          type: integer
          minimum: 1
          description: Bet amount in cents.
        predictionType:
          type: string
          enum: [over, under, exact]
        predictionValue:
          type: integer
          minimum: 1
          maximum: 6
          description: Required when predictionType is "exact".
      required:
        - betAmount
        - predictionType
    DiceRollResponse:
      type: object
      properties:
        roll:
          type: integer
        won:
          type: boolean
        payout:
          type: integer
          description: Total credited back in cents; 0 on a loss.
        balanceAfter:
          type: integer
        serverSeedHash:
          type: string
        clientSeed:
          type: string
        nonce:
          type: integer
      required:
        - roll
        - won
        - payout
        - balanceAfter
        - serverSeedHash
        - clientSeed
        - nonce
    RouletteSpinRequest:
      type: object
      properties:
        betAmount:
          type: integer
          minimum: 1
          description: Bet amount in cents.
        betType:
          type: string
          enum: [straight, color, parity, range, dozen]
        betValue:
          type: string
          description: >
            Meaning depends on betType — straight: "0".."36"; color: "red"|"black";
            parity: "even"|"odd"; range: "low"|"high"; dozen: "1st"|"2nd"|"3rd".
      required:
        - betAmount
        - betType
        - betValue
    RouletteSpinResponse:
      type: object
      properties:
        number:
          type: integer
        won:
          type: boolean
        payout:
          type: integer
          description: Total credited back in cents; 0 on a loss.
        balanceAfter:
          type: integer
        serverSeedHash:
          type: string
        clientSeed:
          type: string
        nonce:
          type: integer
      required:
        - number
        - won
        - payout
        - balanceAfter
        - serverSeedHash
        - clientSeed
        - nonce
    SlotsSpinRequest:
      type: object
      properties:
        betAmount:
          type: integer
          minimum: 1
          description: Bet amount in cents.
      required:
        - betAmount
    SlotsSpinResponse:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
          description: The 5 reel results, in order.
        won:
          type: boolean
        multiplier:
          type: number
        payout:
          type: integer
          description: Total credited back in cents; 0 on a loss.
        balanceAfter:
          type: integer
        serverSeedHash:
          type: string
        clientSeed:
          type: string
        nonce:
          type: integer
      required:
        - symbols
        - won
        - multiplier
        - payout
        - balanceAfter
        - serverSeedHash
        - clientSeed
        - nonce
    LuckyPickDrawRequest:
      type: object
      properties:
        betAmount:
          type: integer
          minimum: 1
          description: Bet amount in cents.
        betType:
          type: string
          enum: [single, jodi, patti]
        betSide:
          type: string
          enum: [open, close]
          description: Required for "single" and "patti"; ignored for "jodi".
        pick:
          type: string
          description: >
            Meaning depends on betType — single: "0".."9"; jodi: "00".."99";
            patti: a 3-digit non-decreasing string, e.g. "025".
      required:
        - betAmount
        - betType
        - pick
    LuckyPickDrawResponse:
      type: object
      properties:
        openDigits:
          type: array
          items:
            type: integer
        openAnk:
          type: integer
        openPatti:
          type: string
        closeDigits:
          type: array
          items:
            type: integer
        closeAnk:
          type: integer
        closePatti:
          type: string
        jodi:
          type: string
        won:
          type: boolean
        payout:
          type: integer
          description: Total credited back in cents; 0 on a loss.
        balanceAfter:
          type: integer
        serverSeedHash:
          type: string
        clientSeed:
          type: string
        nonce:
          type: integer
      required:
        - openDigits
        - openAnk
        - openPatti
        - closeDigits
        - closeAnk
        - closePatti
        - jodi
        - won
        - payout
        - balanceAfter
        - serverSeedHash
        - clientSeed
        - nonce
...