# lib/api-zod/src/generated/types/diceRollResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiceRollResponse {
  roll: number;
  won: boolean;
  /** Total credited back in cents; 0 on a loss. */
  payout: number;
  balanceAfter: number;
  serverSeedHash: string;
  clientSeed: string;
  nonce: number;
}