# lib/api-zod/src/generated/types/rotateSeedResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RotateSeedResult {
  /** The raw seed that was active until now (null if there was no prior seed to reveal). */
  revealedServerSeed?: string | null;
  revealedServerSeedHash?: string | null;
  newServerSeedHash: string;
  clientSeed: string;
}