# lib/api-zod/src/generated/types/rouletteSpinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouletteSpinRequestBetType } from "./rouletteSpinRequestBetType";

export interface RouletteSpinRequest {
  /**
   * Bet amount in cents.
   * @minimum 1
   */
  betAmount: number;
  betType: RouletteSpinRequestBetType;
  /** Meaning depends on betType — straight: "0".."36"; color: "red"|"black"; parity: "even"|"odd"; range: "low"|"high"; dozen: "1st"|"2nd"|"3rd".
   */
  betValue: string;
}