# lib/api-zod/src/generated/types/slotsSpinRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SlotsSpinRequest {
  /**
   * Bet amount in cents.
   * @minimum 1
   */
  betAmount: number;
}